+++ graphic.h
@@ -8,6 +8,6 @@
 enum Shape{Circle, Line, Square};
 enum Color{Green, Blue, Black, Red, Grey};
 
-void drawEntity(Shape shape,Color color, int x, int y, int size, double angle=0.);
+void drawEntity(Shape shape,Color color, double x, double y, double size, double angle=0.);
 
 #endif //MICRORECIF_GRAPHIC_H
